+++ src/main/java/com/vonage/client/verify/ControlResponse.java
@@ -46,6 +46,11 @@ public class ControlResponse {
         return command;
     }
 
+    /**
+     * If the status is non-zero, this explains the error encountered.
+     *
+     * @return The error description, or <code>null</code> if there was no error.
+     */
     @JsonProperty("error_text")
     public String getErrorText() {
         return errorText;