+++ src/main/java/com/vonage/client/verify/SearchRequest.java
@@ -25,9 +25,10 @@ public class SearchRequest {
 
     public SearchRequest(final String... requestIds) {
         if (requestIds.length == 0) {
-            throw new IllegalArgumentException("At least one SnsRequest ID must be provided in a SearchRequest");
-        } else if (requestIds.length > MAX_SEARCH_REQUESTS) {
-                throw new IllegalArgumentException("too many request IDs. Max is " + MAX_SEARCH_REQUESTS);
+            throw new IllegalArgumentException("At least one request ID must be provided in a SearchRequest");
+        }
+        else if (requestIds.length > MAX_SEARCH_REQUESTS) {
+            throw new IllegalArgumentException("too many request IDs. Max is " + MAX_SEARCH_REQUESTS);
         }
         this.requestIds = requestIds;
     }