+++ src/main/java/com/vonage/client/verify/VerifyResponse.java
@@ -16,6 +16,7 @@
 package com.vonage.client.verify;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -23,8 +24,9 @@ import com.vonage.client.VonageResponseParseException;
 import com.vonage.client.VonageUnexpectedException;
 import java.io.IOException;
 
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class VerifyResponse {
-    private VerifyStatus status;
+    private final VerifyStatus status;
     private String requestId, errorText, network;
 
     @JsonCreator
@@ -32,20 +34,34 @@ public class VerifyResponse {
         this.status = status;
     }
 
+    /**
+     * @return The unique ID of the Verify request. You need this <code>request_id</code> for the Verify check.
+     * Note that this may not be present when {@link #getNetwork()} returns a non-null / non-empty value.
+     */
     @JsonProperty("request_id")
     public String getRequestId() {
         return requestId;
     }
 
+    /**
+     * @return The outcome of the request; <code>0</code> (i.e. {@link VerifyStatus#OK}) indicates success.
+     */
     public VerifyStatus getStatus() {
         return status;
     }
 
+    /**
+     * @return If status is non-zero, this explains the error encountered.
+     */
     @JsonProperty("error_text")
     public String getErrorText() {
         return errorText;
     }
 
+    /**
+     * @return The network ID, if {@link #getStatus()} returns {@link VerifyStatus#NUMBER_BARRED}.
+     * @since 7.1.0
+     */
     @JsonProperty("network")
     public String getNetwork() {
         return network;