+++ src/test/java/com/vonage/client/verify/VerifyClientSearchEndpointTest.java
@@ -17,13 +17,13 @@ package com.vonage.client.verify;
 
 import com.vonage.client.ClientTest;
 import com.vonage.client.VonageResponseParseException;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import org.junit.Before;
 import org.junit.Test;
 import java.math.BigDecimal;
 import java.text.SimpleDateFormat;
 import java.util.List;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
 
 public class VerifyClientSearchEndpointTest extends ClientTest<VerifyClient> {
     @Before
@@ -91,12 +91,12 @@ public class VerifyClientSearchEndpointTest extends ClientTest<VerifyClient> {
     @Test
     public void testSearchError() throws Exception {
         //language=JSON
-        String json = "{\n" + "  \"request_id\": \"\",\n" + "  \"status\": \"101\",\n"
-                + "  \"error_text\": \"No response found.\"\n" + "}";
+        String json = "{\n" + "  \"request_id\": \"\",\n" + "  \"status\": \"16\",\n"
+                + "  \"error_text\": \"The code inserted does not match the expected value.\"\n" + "}";
         wrapper.setHttpClient(stubHttpClient(200, json));
 
         SearchVerifyResponse response = client.search("AAAAA");
-        assertEquals(VerifyStatus.NO_RESPONSE, response.getStatus());
+        assertEquals(VerifyStatus.INVALID_CODE, response.getStatus());
     }
 
     @Test
@@ -226,6 +226,40 @@ public class VerifyClientSearchEndpointTest extends ClientTest<VerifyClient> {
         assertEquals("another-random-request-id", second.getRequestId());
         assertEquals("verify2", second.getSenderId());
         assertEquals(VerifyDetails.Status.EXPIRED, second.getStatus());
+    }
+
+    @Test
+    public void testSearchMultipleRequests() throws Exception {
+        String accountId = "abcde";
+        String json = "{\"verification_requests\":[{" +
+                "\"request_id\":\"request-id-1\"," +
+                "\"account_id\": \"abcde1\"},{" +
+                "\"request_id\":\"request-id-2\"," +
+                "\"account_id\": \"abcde2\"},{" +
+                "\"request_id\":\"request-id-3\"," +
+                "\"account_id\": \"abcde3\"}]}";
+
+        wrapper.setHttpClient(stubHttpClient(200, json));
+        SearchVerifyResponse response = client.search("request-id-1", "request-id-2", "request-id-3");
+        List<VerifyDetails> requests = response.getVerificationRequests();
+        assertEquals(3, requests.size());
+        assertEquals(requests.get(0).getAccountId(), accountId+"1");
+        assertEquals(requests.get(1).getAccountId(), accountId+"2");
+        assertEquals(requests.get(2).getAccountId(), accountId+"3");
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void testNoSearchRequests() throws Exception {
+        client.search();
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void testTooManySearchRequests() throws Exception {
+        String[] requestIds = new String[11];
+        for (int i = 1; i <= requestIds.length; i++) {
+            requestIds[i-1] = "request-id-"+i;
+        }
+        client.search(requestIds);
     }
 
     @Test(expected = VonageResponseParseException.class)